third_party/blink/web_tests/platform/generic/fast/forms/lazy-event-listener-scope-chain-expected.txt

This test tests that a lazy event listener attached to a form element keeps its form in the scope chain when the listener is called by JavaScript.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is "abc"
PASS result is "abc"
PASS result is "abc"
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE

 
